requirements.txt
streamlit
yfinance
pandas
numpy
plotly
matplotlib
reportlab
openai
python-dotenv
fredapi
requests